requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==1.3.0
async-generator==1.10
attrs==22.2.0
behave==1.2.6
Brotli==1.0.9
browserstack-local==1.2.7
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==38.0.4
dill==0.3.8
distlib==0.3.8
docutils==0.18.1
exceptiongroup==1.1.1
execnet==1.9.0
filelock==3.14.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jsonmerge==1.9.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
multiprocess==0.70.16
numpy==1.26.0
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.0
parse==1.19.0
parse-type==0.6.0
Paver==1.3.4
Pillow==10.0.0
platformdirs==2.6.2
pluggy==1.5.0
psutil==5.9.8
py==1.11.0
pycparser==2.22
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.1.3
pytest-variables==3.1.0
pytest-xdist==3.3.1
Python-BrowserStack==1.2
PyYAML==6.0
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
selenium==3.141.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tornado==6.3.3
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
utils==1.0.2
virtualenv==20.15.1
wsproto==1.2.0